Fix rollback() property notifiers
Added keys argument to _notifyVirtualProperties(), so you can pass in
an array of keys instead of always using the changes and error keys.

rollback() has been changed to get the keys first, and then notify the
properties after resetting the changes and errors.

Added _rollbackKeys() private function, so rollback() and
_notifyVirtualProperties() can use the keys in different ways.