Add failing test for getting nested value in error object
In v1.6.0, it was possible to get a nested value from the `error` object by using `get` on an instance of Changeset:

```js
changeset.get('error.name')
```

This returned `undefined` if it wasn't present, and an object with the `value` and `validation` if it was.

This is not possible in the newest version - it instead returns `undefined` no matter what.

By using `git bisect` with this test, I found that this stopped working in `v2.0.0-beta.0`.

I don't know what the intended behavior here is, but it seems to me like this at least _was_ intended to work.

There's an example in the README that uses this too, which also fails since 2.0: https://github.com/poteto/ember-changeset/blob/3d777266a252d352513c6833fe3fc83345da9c90/README.md#tips-and-tricks

Here too, this always returns `undefined` and sets `hasError` to `false`:

```js
    if (!changeset.get(`error.${valuePath}`)) {
      set(this, 'hasError', false);
    } else {
      // if error, restore changeset so don't show error in template immediately'
      // i.e. wait for onblur action to validate and show error in template
      changeset.restore(snapshot);
    }
```