Fix test for sync belongsTo()
In case of sync `profile` relationship user.profile is null after
executing the changeset. For safe check we need to use
`user.get('profile.firstName')` vs
`user.get('profile').get('firstName')`

user.get('profile') is always a proxy in case of async belongsTo(),
and it doesn't correlate with calling `save()` on the user model.
So, this check doesn't really check anything.
For the sync relationship this check fails, because profile is null
after performing `execute()`