Extend test coverage to sync relationships (#409)
* Perform tests with sync relationships

Ember-changeset is supposed to work with sync and async relationships.
Behavior differs significantly, we need to test for both types of
relationships.

Create `sync-user` model which is mirroring the `user` user model but
it uses sync relationships.

Refactor tests for users to be executable for both types of users.

* Fix test for sync belongsTo()

In case of sync `profile` relationship user.profile is null after
executing the changeset. For safe check we need to use
`user.get('profile.firstName')` vs
`user.get('profile').get('firstName')`

user.get('profile') is always a proxy in case of async belongsTo(),
and it doesn't correlate with calling `save()` on the user model.
So, this check doesn't really check anything.
For the sync relationship this check fails, because profile is null
after performing `execute()`

* Add tests for initially empty relationships

* Bandaid fix to merge tests in

Add more heuristics to support sync belongsTo() relationships.
A preferable fix is to let user specify if an object is ED Model or not.
One of the options is passing a flag in the constructor or having a
separate implementation for ED Models.