Failing test case for validating computed property (#482)
* Test case for validating changeset getter

The use case is the following:
I want to be able to validate derived state, not the root state.
So, model contains a bunch of changes, but the condition of validity
is defined based on several changes.

With the regular approach when changes are applied to a model, I can
define a computed property (or a getter) and validate against it.

With the changeset approach, model's computed property won't be updated
because the model itself is not changed. So, the idea is to define a
computed property on the Changeset level. But it isn't even evaluated.

This case has been supported in validated-changeset.
This PR adds a test on the ember-changeset level.

* Bump validated-changeset to 0.5.11

The issue which this PR illustrates was fixed in
validated-changeset@0.5.11