#512 add tests for ember-data models with `todo` mark
This tests shows that currently mergeDeep is not handling merging of models without mutation.
Approach that was takes for POJO objects, allows to mergeDeep into empty object (`{}`) and then model and changes. This way, original data is left untouch.

With ember-data model merging Model into `{}` don't give you a model, but gives a different object.